update add Name when create user
